+++ src/components/book-appointment-cta.jsx
@@ -2,6 +2,7 @@ import Image from "next/image";
 import { Button } from "@/components/ui/button";
 import { MessageCircle } from "lucide-react";
 import Link from "next/link";
+import BookConsultationButton from "./BookConsultationButton";
 
 export function BookAppointmentCTA() {
   return (
@@ -22,16 +23,22 @@ export function BookAppointmentCTA() {
             <p className="text-4xl md:text-5xl font-bold text-custom-green-dark">
               +91-9152807080
             </p>
-
+            <div className="flex flex-col md:flex-row items-center justify-center gap-8">
             <Link
               href="https://wa.me/9152807080"
               target="_blank"
               rel="noopener noreferrer"
-              className="bg-green-600 hover:bg-custom-green-dark/90 w-full md:w-[25%] md:h-[15%] text-white px-8 py-6 text-lg font-semibold rounded-full shadow-md flex items-center gap-2 mx-auto"
+              className="bg-green-600 hover:bg-custom-green-dark/90 w-full md:w-[25%] md:h-[15%] text-white px-8 py-6 text-lg font-semibold rounded-full shadow-md flex items-center gap-2"
             >
               <MessageCircle className="h-6 w-6" />
               CHAT ON WHATSAPP!
-            </Link>
+              </Link>
+              <BookConsultationButton
+                  buttonText="Book Your Consultation"
+                  className="w-full md:w-[25%] md:h-[15%] rounded-full px-8 py-6 text-base bg-amber-600 text-white"
+                  display="h-8 w-8 block"
+              />
+            </div>
           </div>
         </div>
       </div>
