+++ src/components/footer.jsx
@@ -1,6 +1,7 @@
 import Link from "next/link";
 import { Facebook, Instagram, Mail, MapPin, Phone } from "lucide-react";
 import Image from "next/image";
+import BookConsultationButton from "./BookConsultationButton";
 
 export function Footer() {
   return (
@@ -53,6 +54,7 @@ export function Footer() {
             Results vary based on multiple factors such as gender, age,
             genetics, activity factor & compliance on Diet & other instructions.
           </p>
+          <BookConsultationButton buttonText="Book Your Consultation" className="px-4 py-4 text-base bg-transparent text-gray-800 hover:bg-gray-50 shadow-none ring-1" display="h-6 w-6 block"/>
         </div>
       </div>
     </footer>
